requirements.txt
flask==2.3.3
Flask-Dropzone==1.6.0
Flask-Session==0.5.0
WTForms==3.0.1
Flask-WTF==1.1.1
wtforms-validators==1.0.0
app==0.0.1
form==0.0.1
python-form==0.2.3
secrets==1.0.2
opencv-python==4.8.0.76
cv2-tools==2.4.0
pytesseract==0.3.10
tesseract==0.1.3
Pillow==10.0.1
numpy==1.26.0
googletrans==3.0.0
playsound==1.3.0
gTTS==2.3.2
